refactor: websocket session 관리 class 분리
